+++ spaceship/include/computer.h
@@ -1,12 +1,13 @@
 #pragma once
+#include <cstdint>
 #include <deque>
 #include <vector>
 
 namespace AdventOfCode {
-using Program = std::vector<int>;
-using Memory = std::vector<int>;
-using Input = std::deque<int>;
-using Output = std::deque<int>;
+using Program = std::vector<int64_t>;
+using Memory = std::vector<int64_t>;
+using Input = std::deque<int64_t>;
+using Output = std::deque<int64_t>;
 
 enum class Intcode {
   Add = 1,
@@ -26,28 +27,28 @@ public:
   Computer (Program const& program);
 
   Memory& accessMemory ();
-  int relativeBase () const;
-  void adjustBase (int offset);
+  int64_t relativeBase () const;
+  void adjustBase (int64_t offset);
   
   void calculate();
   Intcode runInstruction ();
 
   void writeInput(Input const &input);
-  int readInput();
+  int64_t readInput();
 
   void writeOutput(Output const &output);
-  int readOutput();
+  int64_t readOutput();
 
-  void moveInstructionPointer(int offset);
-  void setInstructionPointer(int address);
-  int getInstructionPointer() const;
+  void moveInstructionPointer(int64_t offset);
+  void setInstructionPointer(int64_t address);
+  int64_t getInstructionPointer() const;
 
 private:
   Memory memory_;
   Input input_;
   Output output_;
-  int instructionPointer_;
-  int relativeBase_;
+  int64_t instructionPointer_;
+  int64_t relativeBase_;
 };
 
 } // namespace AdventOfCode
